requirements.txt
python-dotenv==1.0.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
colorama>=0.4.6
PySocks>=1.7.1
setuptools>=70.0.0
